Direction/Design Change
We realized it makes more sense to have the controller update at the end of all logic. Also, we're going to move a lot of code out of global variables and into an object. This will allow us to automate more of the controller, but we can still give the end user absolute control through scripts. See the updated txt for details.