Determine GP or KB changes and removed "update"
I've changed the code so it no longer requires an "update". All the scripts are designed to automatically figure out what they need when they're called. Even an init script is no longer needed (technically it is, but all scripts call it if needed). This may be close to a finished version.